+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+- Speed up `Baker.get_fields()` by returning `_meta.fields + _meta.many_to_many` directly instead of filtering `_meta.get_fields()`.
 
 ### Removed